add: share button, attempt to get share to twit working
