+++ package.json
@@ -59,8 +59,8 @@
 		"moment-business-time": "1.1.1",
 		"next": "12.1.6",
 		"next-compose-plugins": "2.2.0",
-		"next-optimized-images": "^2.6.2",
 		"next-connect": "^0.12.2",
+		"next-optimized-images": "^2.6.2",
 		"react": "17.0.1",
 		"react-countdown": "2.3.1",
 		"react-dom": "17.0.1",
@@ -127,6 +127,7 @@
 		"eslint-plugin-react-hooks": "4.2.0",
 		"eslint-plugin-testing-library": "^5.1.0",
 		"eslint-plugin-ui-testing": "^2.0.0",
+		"html-to-image": "^1.9.0",
 		"husky": "5.0.9",
 		"imagemin-optipng": "8.0.0",
 		"imagemin-svgo": "8.0.0",
