+++ sections/futures/ShareModal/ShareModal.tsx
@@ -1,16 +1,16 @@
-import { FC } from 'react';
-import { format } from 'date-fns';
+import { FC, useCallback, useRef, useState } from 'react';
 import styled from 'styled-components';
 import { useTranslation } from 'react-i18next';
+import { toPng } from 'html-to-image';
 
+import Button from 'components/Button';
 import BaseModal from 'components/BaseModal';
-import CurrencyIcon from 'components/Currency/CurrencyIcon';
 
 import { CurrencyKey } from 'constants/currency';
-import getLocale from 'utils/formatters/getLocale';
-import { formatNumber, zeroBN } from 'utils/formatters/number';
-import { FuturesPosition, PositionHistory, PositionSide } from 'queries/futures/types';
+import { FuturesPosition, PositionHistory } from 'queries/futures/types';
 
+import AmountContainer from './AmountContainer';
+import PositionMetadata from './PositionMetadata';
 import PNLGraphicPNG from 'assets/png/pnl-graphic.png';
 
 type ShareModalProps = {
@@ -21,30 +21,6 @@ type ShareModalProps = {
 	futuresPositionHistory: PositionHistory[];
 };
 
-const lineSeparatorStyle = { margin: '0px 0.7vw 0px 0.7vw' };
-const currencyIconStyle = {
-	height: '1.94vw',
-	width: 'auto',
-	margin: '-0.3vw 0.5vw 0vw 0vw',
-};
-
-const getFontFamily = (props: any) => {
-	/**
-	 * @todo `open-at` and `close-at` must use the `GT America` font-family
-	 */
-	if (props.className === 'open-at-header' || props.className === 'created-on-header')
-		return props.theme.fonts.regular;
-	if (
-		props.className === 'open-at-date' ||
-		props.className === 'avg-open-price' ||
-		props.className === 'created-on-date' ||
-		props.className === 'current-price'
-	)
-		return props.theme.fonts.monoBold;
-	if (props.className === 'open-at-time' || props.className === 'created-on-time')
-		return props.theme.fonts.mono;
-};
-
 const ShareModal: FC<ShareModalProps> = ({
 	position,
 	marketAsset,
@@ -52,40 +28,38 @@ const ShareModal: FC<ShareModalProps> = ({
 	setOpenShareModal,
 	futuresPositionHistory,
 }) => {
-	let avgEntryPrice: string = '',
-		openAtDate: string = '',
-		openAtTime: string = '',
-		createdOnDate: string = '',
-		createdOnTime: string = '';
-
-	if (futuresPositionHistory.length > 0) {
-		avgEntryPrice = futuresPositionHistory[0]?.avgEntryPrice.toNumber().toFixed(2);
-
-		const openTimestamp = futuresPositionHistory[0]?.openTimestamp;
-		const closeTimestamp = futuresPositionHistory[0]?.timestamp;
-
-		openAtDate = format(openTimestamp, 'PP', { locale: getLocale() });
-		openAtTime = format(openTimestamp, 'HH:mm:ss', { locale: getLocale() });
-		createdOnDate = format(closeTimestamp, 'PP', { locale: getLocale() });
-		createdOnTime = format(closeTimestamp, 'HH:mm:ss', { locale: getLocale() });
-	}
-
 	const { t } = useTranslation();
-	const positionDetails = position?.position ?? null;
-	const leverage = formatNumber(positionDetails?.leverage ?? zeroBN) + 'x';
-	const side = positionDetails?.side === 'long' ? PositionSide.LONG : PositionSide.SHORT;
-	const roiChange = positionDetails?.roiChange.mul(100);
-	const marketAsset__RemovedSChar = marketAsset[0] === 's' ? marketAsset.slice(1) : marketAsset;
+	const ref = useRef<HTMLDivElement>(null);
+	const [imagePathName, setImagePathName] = useState<string>('');
 
-	const amount = () => {
-		if (roiChange) {
-			return roiChange.gt(0)
-				? `+${roiChange.toNumber().toFixed(2)}%`
-				: roiChange.eq(0)
-				? `+0.00%`
-				: `${roiChange.toNumber().toFixed(2)}%`;
+	const positionDetails = position?.position ?? null;
+
+	const onDownloadImage = useCallback(() => {
+		if (ref.current === null) {
+			return;
 		}
-	};
+
+		toPng(ref.current, { cacheBust: true })
+			.then((dataUrl: any) => {
+				const link = document.createElement('a');
+
+				link.download = 'my-image-name.png';
+				link.href = dataUrl;
+
+				link.click();
+
+				const pathName = link.pathname;
+				setImagePathName(pathName);
+				console.log('pathName', pathName);
+			})
+			.catch((err: any) => {
+				console.log(err);
+			});
+
+		/**
+		 * @todo Share to twitter and other socials
+		 */
+	}, [ref]);
 
 	return (
 		<>
@@ -95,7 +69,7 @@ const ShareModal: FC<ShareModalProps> = ({
 				title={t('futures.modals.share.title')}
 			>
 				<ModalWindow>
-					<PNLGraphic>
+					<PNLGraphic ref={ref}>
 						<div
 							style={{
 								position: 'relative',
@@ -124,169 +98,36 @@ const ShareModal: FC<ShareModalProps> = ({
 								}}
 							/>
 						</div>
-						<AmountContainer>
-							<StyledPositionType>
-								<CurrencyIcon
-									style={currencyIconStyle}
-									currencyKey={(marketAsset[0] !== 's' ? 's' : '') + marketAsset}
-								/>
-								<StyledPositionDetails>{`${marketAsset__RemovedSChar}-PERP`}</StyledPositionDetails>
-								<StyledPositionDetails style={lineSeparatorStyle}>{`|`}</StyledPositionDetails>
-								<StyledPositionSide className={side}>{side.toUpperCase()}</StyledPositionSide>
-								<StyledPositionDetails style={lineSeparatorStyle}>{`|`}</StyledPositionDetails>
-								<StyledPositionLeverage>{`${leverage}`}</StyledPositionLeverage>
-							</StyledPositionType>
-							<StyledAmount className={side}>{amount()}</StyledAmount>
-						</AmountContainer>
-						<TopLeftContainer>
-							<ContainerText className="open-at-header">{`OPEN AT`}</ContainerText>
-							<ContainerText
-								className="open-at-date"
-								style={{ fontSize: '1.13vw', color: '#FFFF' }}
-							>
-								{openAtDate.toUpperCase()}
-							</ContainerText>
-							<ContainerText
-								className="open-at-time"
-								style={{ fontSize: '0.73vw', letterSpacing: '0.7px' }}
-							>
-								{openAtTime}
-							</ContainerText>
-						</TopLeftContainer>
-						<TopRightContainer>
-							<ContainerText className="created-on-header">{`CREATED ON`}</ContainerText>
-							<ContainerText
-								className="created-on-date"
-								style={{ fontSize: '1.13vw', color: '#FFFF' }}
-							>
-								{createdOnDate.toUpperCase()}
-							</ContainerText>
-							<ContainerText
-								className="created-on-time"
-								style={{ fontSize: '0.73vw', letterSpacing: '0.7px' }}
-							>
-								{createdOnTime}
-							</ContainerText>
-						</TopRightContainer>
-						<BottomLeftContainer>
-							<ContainerText>{`AVG OPEN PRICE`}</ContainerText>
-							<ContainerText
-								className="avg-open-price"
-								style={{ fontSize: '1.13vw', color: '#FFFF' }}
-							>
-								{avgEntryPrice}
-							</ContainerText>
-						</BottomLeftContainer>
-						<BottomRightContainer>
-							<ContainerText>{`CURRENT PRICE`}</ContainerText>
-							<ContainerText
-								className="current-price"
-								style={{ fontSize: '1.13vw', color: '#FFFF' }}
-							>
-								{marketAssetRate.toFixed(2)}
-							</ContainerText>
-						</BottomRightContainer>
+						<AmountContainer marketAsset={marketAsset} position={position} />
+						<PositionMetadata
+							marketAssetRate={marketAssetRate}
+							futuresPositionHistory={futuresPositionHistory}
+						/>
 					</PNLGraphic>
+					<ButtonContainer>
+						<Button
+							variant="primary"
+							isRounded={true}
+							onClick={onDownloadImage}
+							size="sm"
+							disabled={false}
+						>
+							{t('futures.modals.share.button')}
+						</Button>
+					</ButtonContainer>
 				</ModalWindow>
 			</BaseModal>
 		</>
 	);
 };
 
-const ContainerText = styled.div`
-	font-size: 0.56vw;
-	color: #999999;
-
-	font-family: ${(props) => getFontFamily(props)};
-`;
-
-const TopRightContainer = styled.div`
-	position: absolute;
-	display: flex;
-	flex-direction: column;
-
-	bottom: 5.5vw;
-	left: 12.02vw;
-`;
-
-const TopLeftContainer = styled.div`
-	position: absolute;
-	display: flex;
-	flex-direction: column;
-
-	bottom: 5.5vw;
-	left: 2.02vw;
-`;
-
-const BottomRightContainer = styled.div`
-	position: absolute;
-	display: flex;
-	flex-direction: column;
-
-	bottom: 2vw;
-	left: 12.02vw;
-`;
-
-const BottomLeftContainer = styled.div`
-	position: absolute;
-	display: flex;
-	flex-direction: column;
-
-	bottom: 2vw;
-	left: 2.02vw;
-`;
-
-/** @todo Import and use correct fonts from Figma design */
-const StyledPositionLeverage = styled.div`
-	display: flex;
-	flex-direction: column;
-	color: #c9975b;
-
-	font-size: 1.07vw;
-`;
-
-const StyledPositionSide = styled.div`
-	display: flex;
-	flex-direction: column;
-	color: ${(props) => (props.className === 'long' ? '#7fd482' : '#ff0420')};
-
-	font-size: 1.07vw;
-`;
-
-const StyledPositionDetails = styled.div`
-	display: flex;
-	flex-direction: column;
-
-	font-size: 1.07vw;
-
-	color: #ffff;
-`;
-
-const StyledPositionType = styled.div`
+const ButtonContainer = styled.div`
 	display: flex;
 	flex-direction: row;
-`;
 
-/** @todo Import and use correct fonts from Figma design */
-const StyledAmount = styled.div`
-	position: absolute;
-	margin-top: -0.05vw;
+	margin: 25px 0px 25px 0px;
 
-	font-size: 4.8vw;
-	font-weight: 700;
-	color: ${(props) => (props.className && parseFloat(props.className) > 0 ? '#7fd482' : '#ff0420')};
-
-	text-shadow: 0px 0px 3.99vw
-		${(props) =>
-			props.className && parseFloat(props.className) > 0
-				? 'rgba(127, 212, 130, 0.35)'
-				: 'rgb(255, 4, 32, 0.35)'};
-`;
-
-const AmountContainer = styled.div`
-	position: absolute;
-	top: 6vw;
-	left: 2.02vw;
+	justify-content: center;
 `;
 
 const PNLGraphic = styled.div`
