+++ translations/en.json
@@ -896,9 +896,7 @@
 			},
 			"share": {
 				"title": "Share",
-				"something": {
-					"something-else": "some text"
-				}
+				"button": "Share"
 			}
 		}
 	},
